move test into retry_test.go, deflake picker, drop mutex
